RemoteAcquisitionCLI
No warranty on the usage of this product. Be sure to unzip the dependency support tools for any of the scripts to work. Deploy the scripts and binaries on windows boxes through tools like SCCM or PSexec and you can acquire logical, physical, and memory images to a remote share.